Add TalentLink RAC JavaScript app
SABA TalentLink's Responsive Application Component is a JavaScript
application to handle user job application forms, and other
functionality like uploading CVs and saving and returning to a partial
application.

This commit inserts the HTML code, CSS and JavaScript files almost
exactly as provided by SABA TalentLink, and routes to the application
using a route on RecruitmentHomePage.

We insert linting exemption lines for all three static files, as they're
not under our control.

The HTML file was supplied as a whole page. I have left the original
content commented out, and inserted Django template block tags to make
it work with the recruitment site. This is intended to make it easier to
update the template in future if a new HTML file is supplied in a
similar fashion.

The CSS in this commit corresponds to the old site's styling. New styles
are to follow.