Update Elasticsearch settings, and add synonyms
Add the search.Term model Enable synonym lookup, configurable in the
admin (the synonym filter is applied at index time, so changes do not
happen instantly).
Set the Porter stemming algorithm filter, which is designed for English
grammar (e.g. married = marry => marri)
Add an English stop words filter, to reduce the number of hits for words
like and, of, the.