Bump Wagtail from 2.9.3 to 2.10.2
This is the main objective of this branch. Upgrade to 2.10 is needed to
enable the use of `wagtail-automatic-redirect` with the `post_page_move`
signal to automatically create redirects when a page is moved.

The `post_page_move` signal was only added to Wagtail in 2.10. Without
the upgrade to 2.10, `wagtail-automatic-redirects` can only create
redirects for changes of page slugs.

Changing to Wagtail 2.10 should unblock the automatic redirect ticket:
https://trello.com/c/p7SaAQTP/467-implement-wagtail-automatic-redirect