Add `clean_field` migrations and to fixture
Wagtail 2.10 adds a `clean_field` property to the `AbstractFormField`.
`bc.forms.FormField` inherits from `AbstractFormField`.

This commit adds the migration to add the `clean_name` column for
`FormField` entries.

This commit also adds the `clean_field` to the fixtures. The tests did
not fail without the data added, but I think that is only because the
fixtures are currently not used in any test. To avoid future failure
I have added the field now, while the context is clear.

For more information on the `clean_field` see the Wagtail 2.10 release
notes:
https://docs.wagtail.io/en/stable/releases/2.10.html#clean-name-field-added-to-form-builder-form-field-models