Add test for IndexPage under FIS CategoryType page
Before, it was only tested that it is possible to add the `IndexPage`
under `HomePage` or itself.

It was specifically requested that it can also added to the FIS category
type pages (one and two).

This commit changes the respective tests.

Some discussion might be needed why the test
`test_can_only_create_indexpage_under_homepage` was limiting the
possible parents explicitly. The issue was raised on the ticket:
https://trello.com/c/tEdgGiDw/475-add-functionality-for-implementing-row-of-3-boxes-from-main-site-to-fis-microsite-index-pages#comment-5fcfef4ed1fae9282e95eaa4

The test was changed to be able to move forward for now.