Add FIS cat pages as possible IndexPage parents
Before, it was not possible to add `IndexPage` as a child to the two
FIS category type pages (one and two). The previously defined tests for
that behaviour failed.

This commit allows for the FIS category pages to serve as parents for
the `IndexPage`.

This behaviour was specifically requested and is the goal of the ticket.
But, it violates a previously defined test. The test was changed in the
last commit to allow for the requested functionality, but a further
discussion is needed.

See also: https://trello.com/c/tEdgGiDw/475-add-functionality-for-implementing-row-of-3-boxes-from-main-site-to-fis-microsite-index-pages#comment-5fcfef4ed1fae9282e95eaa4