Add IndexPage as subpage type for BaseCatPage
Before, when trying to add a child page to either FIS category page,
it would automatically add an `InlineIndex`.

This was because that was the only page type listed in
`BaseCategoryPage.subpage_types`. Both `CategoryTypeOnePage` and
`CategoryTypeTwoPage` inherit from this page type (or rather implement
it -- since `BaseCategoryPage` is an abstract class that can not be
instantiated itself).

This commit adds `standardpages.IndexPage` to
`BaseCategoryPage.subpage_types` which enabled the desired behaviour in
the admin.

Now, when editors click to add a child page to `CategoryTypeOnePage` or
`CategoryTypeTwoPage` they are asked if the new child should be of type
`InlineIndex` or `IndexPage`.