Allow InformationPage as child for FIS cat 1 & 2
Before, it was only possible to add `IndexPage` and `InlineIndex` as
child pages for the FIS `CategoryTypeOnePage` or `CategoryTypeTwoPage`.

This commit extends the list to include `InformationPage`.

This is feature is requested in a separate ticket than the addition of
`IndexPage` as a possible child of the category type pages. But, because
the change is so similar and would only create a merge conflict to be
resolved, this functionality is added here.

Ticket: https://trello.com/c/cC21jNOI/476-add-functionality-for-implementing-information-page-from-main-site-to-fis-microsite