Add tests for possible children of category pages
Before, it was tested in the tests for `IndexPage` that the `IndexPage`
could be added  to the FIS category pages. These tests had been added in
the `IndexPage` tests because a similar test already existed to check
the possible creation under `home.HomePage`.

Now the same test for the `InformationPage` was necessary. For the
`InformationPage` is did not make sense to add the test in the its
tests, because there was no change necessary in the `InformationPage`
to add this behaviour.

The change was necessary in the abstract `BaseCategoryPage` which both
category pages implement. Therefore, it seemed more sensible to add the
tests for the category pages rather than the `InformationPage`.

For consistency, the same test for the `IndexPage` has also been moved
to the tests for the category pages. In this case the test location
is not obvious, because the behaviour depends on both models (because
`IndexPage` defines the `parent_page_types` property).