Add test for excluding FIS pages from search
To make sure that the desired behaviour (excluding FIS pages from the
search results) is delivered a test is added. This will also make it
obvious when another change might reactivate the undesired behaviour.

Once the FIS pages are supposed to be included, the only change
necessary is that the one assertion of the test should be inverted.

To make the setup easy, this commit also adds the fixtures for the
FIS homepage and the category one page.