Exclude FIS pages from search source pages
Before, the search would return pages from any site/root page. This
is undesired, because the FIS page in production is not ready for
use.

To avoid preliminary content from appearing in search, the pages that
start with the same path as the FIS homepage are excluded from the
page queryset on which the search is performed.

The exclusion is applied through a single method that is called on the
page queryset. This makes it easy to revert this change in the future
once the FIS content is ready.

See also: https://trello.com/c/TryuPZ9J/478-update-search-configuration-to-exclude-fis-content