Add test scaffolding example
This took for ever to figure out how to just setup a normal InlineIndex
page. For some reason it would never be able to reach the url of the
InlineIndex page. Without creating a site object the url property was
empty. But then then trying to access the page with the test client
still resulted in 404...

I have no idea why.

Searching for other use cases of the test client pointed me to trying
to use the `HomePage.objects.first()` to retrieve the root homepage to
which I can add the inline index page. This works but I am not sure why.

Trying to access the homepages that I had been creating before did
work with the test client. But the inline index page would not work...