Add tests for desired display functionality
This adds the proper test to check if the the desired functionality
is reached. This currently fails because draft child pages are listed
on live requests to the live index. This was expected and is the bug
that needs to be solved.

I am testing draft requests (URL generated as reverse fro the wagtail
admin view for the page that is draft) because that was easier to setup
then preview requests. The rendered page should be the same.