Add method for filtered children
This commit adds a method to the `InlineIndex` that returns the children
of the `InlineIndex` with respect to a `show_draft_children` property.
This property is populated during the `get_context` call.

The get context call seemed like the only viable place, because that is
where the request is available (which I need to determine if I want to
show the draft children).

This solution works for the requests to the inline index so far... But
I am not really liking it. Seems dirty. I think it might be cleaner to
remove the method calls from the template.