Add index, next and prev page to context
Before, the template was calling the methods on the page to get access
to the index and the pages to link to for pagination.

Is I want additional filtering and logic applied to select these page,
I need to pass flags to the template and then pass the flags from the
template back to the methods.

This seems unnecessary because I can perform the manipulation before
hand and only pass the required data to the template.

This commit does not change the logic of what is displayed in the
template it only switches from the method calls to the context
variables.