Draft children in index only on draft request
Before, draft children would be in the index even when live viewing a
page.

The above behaviour is not desired and now prevented.

There is still a reference to the draft page though. Not in the index
but in the pagination link. That also needs to be prevented.

Therefore, tests are still failing.