Fix and increase test specificity
Mainly adds the count check for title occurrences of child and sibling
pages. Before only the presence was tested. But, when the page is
expected to be in the index and in the pagination then two links should
be present. This commit fixes that.

Also adds one tests for prev sibling that had been missed before.

Fixes the test for live child listed on live index. Before this was
testing for the wrong title but still passed. The test is now fixed and
still passing.