Override settings in test for draft requests in CI
On the CI, the tests are run with production settings. This leads to the
draft requests failing with 400.

I did find an the original example for draft requests in
`bc/standardpages/tests/tests.py`. There the setting override is used to
allow "localhost" and "testserver" as hosts.

I have copied that approach and hope that this will allow the testing
on CI to succeed.

(Local testing was and is working.)