Move draft check functions into mixin
Before, the functions to check if a page is in draft and the request is
viewing a draft where defined as module wide functions. This worked but
seemed a bit loosely tied to the usecase.

The functions where turned into methods on a new mixin. This mixin is
applied to the inline index page classes.