Reverse boolean checks to avoid potential db query
I am not quite sure, but using `Page.live` and
`Page.has_unpublished_changes` might trigger a database query.

To avoid this, the order of the boolean checks was changed so that the
potential database query is only done when needed (we already know the
request is at least a preview request).

See also review comment:
https://git.torchbox.com/buckinghamshire-council/bc/-/merge_requests/371#note_133691