Add wagtail-transfer settings
The added settings control which models should be updated when ever a
page related to them is updated. Typically, these related models would
only be created on their first discovery, but would not be updated when
the page is updated.

Also, adding a setting so that users between the instances are matched
by username. I had experienced transfer issues with user accounts that
where created on both instances independently, with the same username.
Since this seemed like something that might happen between the staging
and production environment, I added the `username` as a lookup field for
the `user` model. This means that instead of trying to find the matching
instance via a UUID between the instances, the username can be used.
This is helpful in this case, because one can assume the same username
to mean the same person even if the creation date and id of the user
account are different between staging and production.