Remove utils.relatedpage from update list
Adding this model to the list of models to update with it's parent page
is not necessary. The update already works without this being listed.

I am not quite sure why. Maybe it has something to do with being a
child class of `Orderable` (but that is not mentioned in the Wagtail
Transfer docs).