Exclude PostcodeLookupResponse from transfer
The `PostcodeLookupResponse` model contains the `postcodes` field, which
is of the `django.contrib.postgres.fields.ArrayField` type. This is a
special field for Postgres databases.

Wagtail-Transfer is missing a serializer for this field type. This leads
to error messages like the following when trying to transfer the
`PostcodeLookupResponse` model:
```
...
[23/Dec/2020 21:19:54] "GET /admin/wagtail-transfer/api/check_uid/?uid=0cd819d6-bcc4-5344-b75f-d838f8bdce23 HTTP/1.1" 200 0
[23/Dec/2020 21:19:55] "GET /admin/wagtail-transfer/api/chooser-proxy/source/348/?fields=descendants HTTP/1.1" 200 2774
[2020-12-23 21:20:04,225][5707][ERROR][django.request] Internal Server Error: /admin/wagtail-transfer/import/
Traceback (most recent call last):
  File "/home/vagrant/.virtualenvs/bc/lib/python3.6/site-packages/django/db/backends/utils.py", line 84, in _execute
    return self.cursor.execute(sql, params)
psycopg2.DataError: malformed array literal: "["HP20 1UY"]"
LINE 1: ...n_text" = 'What to do in lockdown', "postcodes" = '["HP20 1U...
                                                             ^
DETAIL:  "[" must introduce explicitly-specified array dimensions.
```

To prevent Wagtail-Transfer from attempting to transfer this model (when
found through a relationship to a page) it is added to the
`WAGTAILTRANSFER_NO_FOLLOW_MODELS` list in the settings.

This setting is only effective when it is present before the transfer
table is pre-seeded. Otherwise the model is still attempted to be
transferred. To re-pre-seed the transfer tables run the following on the
source instance:
```
$ dj migrate wagtail_transfer zero
$ dj migrate wagtail_transfer
$ dj preseed_transfer_table wagtailcore wagtailimages wagtaildocs alerts images.customimage documents.customdocument events news utils
```

Then create a database backup on the source side and restore it on the
target side to make sure the databases and especially transfer tables
have the same state in the beginning.

When the feature is released to production this requires a reset of the
staging server to the production state anyways. So the above steps are
not needed separately. The above steps need to be performed on staging
(as source) and local (as target) to be able to test the functionality
before releasing to production.