Add run_before in initial homepage migration
Wagtail 2.11 adds a locale field to the Page model, and since the
existing migrations in your project pre-date this, they are designed to
run against a version of the Page model that has no locale field. As a
result, they need to run before the new migrations that have been added
to wagtailcore within Wagtail 2.11. However, in the old version of the
homepage migration, there is nothing to ensure that this sequence is
followed. The actual order chosen is an internal implementation detail
of Django, and in particular is liable to change as you continue
developing your project under Wagtail 2.11 and create new migrations
that depend on the current state of wagtailcore. In this situation, a
user installing your project on a clean database may encounter the
following error when running manage.py migrate:

```
django.db.utils.IntegrityError: NOT NULL constraint failed: wagtailcore_page.locale_id
```

Adding the run_before directive will ensure that the migrations run in
the intended order, avoiding this error.

See also:
https://docs.wagtail.io/en/v2.11.3/releases/2.11.html#run-before-declaration-needed-in-initial-homepage-migration