Rename classname to form_classname in blocks
Basic StreamField block types such as CharBlock previously accepted a
classname keyword argument, to specify a class attribute to appear on
the page editing form.

For consistency with StructBlock, this has now been changed to
form_classname.

The classname argument is still recognised, but deprecated. This commit
ensures that the block definition will work with future Wagtail
releases.