Rm duplicate `legacy.richtext` app after rebase
After rebasing the current feature branch on the release branch, the
`wagtail.contrib.legacy.richtext` app was included twice in the
`INSTALLED_APPS` list.

This led to `django.core.exceptions.ImproperlyConfigured` errors.

One of the additions is now removed. The `dj check` now does not
throw the configuration error anymore.