Fix FIS exclusion from search results after rebase
There were a few conflicts during the rebase. I must have missed to keep
the FIS page exclusion from the search results. Luckily, the test
pointed this out immediately. The method for exclusion was also still
in the code.

The FIS pages are now excluded again from the search results.