Remove `get_queryset` call on search results
Before, after the search results had been retrieved, the result objects
`get_queryset` method was called. This call was only necessary to
exclude the FIS pages from the queryset (which is done with the
querysets `exclude` method).

The exclusion of the FIS pages can be moved to before the search is
even performed. This removes that need to have the search results
available as a queryset again.

This is additionally desirable as the `get_queryset` method only exists
on the `PostgresSearchResults` class, but not on the `BaseSearchResults`
or the `ElasticSearch5SearResults` classes.

Furthermore, the call to `get_queryset` is a duplication of the call the
the Postgres search already performs internally. It is called by the
[`PostgresSearchResults._do_search` method](https://github.com/wagtail/wagtail/blob/14f63519c2c355637cfcb20715971dfaf9346fbf/wagtail/contrib/postgres_search/backend.py#L595)
which is called by the `search` method to populate the
`_search_results_cache` attribute. The `search` method in turn is used
to create the [iterable functionality](https://github.com/wagtail/wagtail/blob/14f63519c2c355637cfcb20715971dfaf9346fbf/wagtail/search/backends/base.py#L240)
of the search results.