Explicitly update search index in tests
Before, when switching to the Elasticsearch backend, I was getting
errors from the Wagtail Search internals because of missing fields in
the Elasticsearch response objects. Also, tests for search tests with
simple lookups where failing (this is nothing that should be affected)
by the annotations.

The fact that simple results could not be looked up properly was a hint
that the search index might be not up to date with the content of the
database.

This commit adds calls to the `update_index` management command (that
is provided by wagtail to update the search index) after the creation
of test pages.

Typically, the index should be updated automatically through signal
handlers provided by Wagtail Search for the `save` and `delete` signals
of all indexed models. For some reason this does not seem to work in the
test setup. Maybe it has something to do with the models being created
through factories.

See also:
https://docs.wagtail.io/en/stable/topics/search/indexing.html#updating-the-index

When using the `update_index` management command to explicitly update
the search index after pages being created, I am not getting the errors
anymore that I was seeing before. There are still tests that are failing
with the Elasticsearch backend that are passing with Postgres. Btw all
tests are still passing when using the Postgres search backend.

Also, this commit also defines different index name for the search index
used during testing and a function to delete the search index after the
tests. For some reason, these solutions do feel a little hacky, but I
do not know of a better way of doing this. I was also trying to kind of
replicate how the Django test runner is handling the database (create
special tables for testing and drop them afterwards), but without the
ability to actually manipulate the test runner.