Add test creating nav settings w cols not links
Add extra tests that creates the site navigation settings with only
columns but not links. This is already passing, but is added for
completeness of test suite.

It also helped figuring out how to test for an empty stream field.
In this case the following check represents the expected result.
```
        self.assertEqual(nav_setting.footer_links.stream_data, [])
```

The same check was added in other tests with the inverse logic
(`assertNotEqual`) to check for the successful saving of the data.