Disable footer columns requirement
Before, in the navigation settings it was required that the footer
columns would be supplied with a value to allow saving of the navigation
settings.

The requirement of the footer columns was not intended. The point of
this ticket is to allow creation and editing of the footer links with
out any footer columns being present.

To disable the requirement, the `StreamBlock` has received the attribute
`required=False` and the `StreamField` the attribute `blank=True`.
Both are necessary due to the current implementation of the
`StreamBlock` to remove the asterisk (*) marker in the admin and
actually allow for the empty field to be submitted.
See [this Wagtail issue](https://github.com/wagtail/wagtail/issues/4306)
for more information.