Add display tests for footer links and columns
It was specifically requested to test that the removal of the columns
requirement does not impact the display of the footer links.
See [this comment on the ticket](https://trello.com/c/E5SOXKDv/464-footer-link-can-not-be-created-edited#comment-5ff874961063e6783fa94872).

Test were added to check that the added footer navigation elements are
present (columns and links together and each alone).