Update cache setting in local settings example
Before, the local settings example defined the `DummyCache` to be used.
Using the `DummyCache` in the local settings caused tests to fail that
do depend on an actual cache to be working.

Now, the example settings file has been updated to use the default
`DatabaseCache`. The tests that were failing with the `DummyCache` are
passing with the `DatabaseCache`.

This should not be an issue on staging or production as these
environments use the actual Redis cache. This change should just prevent
other developers from running into failing tests due to missing cache
(when they are using a local settings file, which is necessary to test
the Elasticsearch feature).