Update get_synonyms function for ES5
During tests is was getting request errors when using the ES5 backend.
```
elasticsearch.exceptions.RequestError: RequestError(400, 'illegal_argument_exception', 'synonym requires either `synonyms` or `synonyms_path` to be configured')
```

Looking into the actual request revealed that `synonyms` was sent, but
was only an empty list. For some reason, the ES5 backend can not deal
with the list being empty.

The function is updated so the list containing the empty string `[""]`
is only returned when the list otherwise would be empty.

This change works with ES5 and ES7.