Bump GOV notify wrapper to 5.7.1
Before, we were using `notifications-python-client` version 5.4.1. After
the last deployment we started to get errors when ever somebody would
try to sign up for the job alert subscription.

https://sentry.io/organizations/torchbox/issues/2152864364/events/?project=1856234

This issue was caused by the `notifications-python-client` which did not
properly limit the version of it's dependency `PyJWT`. During the
deployment `PyJWT` was updated to 2.0.0. The
`notifications-python-client` of version 5.4.1 is not compatible with
that version due to a change of how the tokens are returned.

The compatibility is fixed in 5.7.1 of `notifications-python-client`.
https://github.com/alphagov/notifications-python-client/blob/master/CHANGELOG.md#571

The `PyJWT` dependency still does not have an upper limit defined.
https://github.com/alphagov/notifications-python-client/blob/930e5f78cfd62281059ea878e0034f309453e518/setup.py#L63
To prevent similar issues like this in the future, I am fixing the
version of `PyJWT` in our dependencies.