Update function definition order
Before, the there was no apparent system to the order of the functions
in `bc.search.tests.utils`.

This has been updated to follow the the order required by python scripts
in that the functions needs to be defined (on top) before used. I guess
because the module was not run directly, the order did not cause issues
before.

The functions are also grouped more sensibly now.