Allow 2 SQL queries when using Elasticsearch
Before, the test checking for the number of DB queries used to access
the search results was failing when run with an Elasticsearch backend.

The test was failing because in stead of the expected 1 query it needed
2 queries. Where is difference is coming from is not quite obvious to
me, but it seems like an implementation detail.

According to @nicks, the intent of the test is to check that the results
can be accessed with O(1) DB queries, not necessarily 1. The main point
is that is not O(n).

This is still fulfilled after allowing 2 DB queries when using the
Elasticsearch backend.