Move is_news_page templatetag to news app
It does not really make much sense to have this template tag in the
styleguide. I was confused by the search tags in the styleguide.

I guess typically you would only want the overrides for the pattern
library to be stored there.

It might make sense to move the search templatetags out of that app as
well.