Use <time> tag for news date label (accessibility)
Before, the label was only using a generic `<span>` tag. This works
visually. This does have the downside that it might not be interpreted
correctly by screen readers and other accessibility tools.

Using the `<time>` tag with the `datetime` attribute should increase the
machine readability of the element. The visual representation is not
affected by this change.