Allow InformationPage as FIS home page child
Before, it was only possible to add pages of Category type 1 and 2 as
children.

The point of the ticket is to allow info pages also to be direct
children.

This commit allows add to add information pages.

Currently, these pages are not displayed on the home page, but are only
accessible through their URL directly.