Bump browser-sync to version ^2.26.8
Before, we were running 2.26.7. For some reason, it listed
`ua-parser-js` 0.7.17 in its `package-lock.json`.

`ua-parser-js` < 0.7.20 is licensed with GPL-2.0. We do not want to put
the GPL limitations on this project and therefore need to remove all
dependencies on GPL software.

`browser-sync` with versions >2.26.8 removes `ua-parser-js` from it's
lock file. This remove the GPL barrier due to `ua-parser-js` < 0.7.20.