Upgrade django-basic-auth-ip-whitelist
We were experiencing issues with Cloudfront caching 401 responses. This
led to people being unable to access the staging site that is protected
by basic auth.

To avoid this, an update was made in `django-basic-auth-ip-whitelist`
to send the "never cache" header on pages protected with basic auth.
This update has not been released in a new version. Therefore, the fixed
version of `django-basic-auth-ip-whitelist` is installed directly from
Github.

See also:
https://github.com/tm-kn/django-basic-auth-ip-whitelist/pull/6