Add Github actions workflow to publish docs
Currently, the docs are build in on the Gitlab CI and published on Gitlab Pages.
The Gitlab pages are login protected and only accessible to Torchbox devs.

When the project becomes open source, it makes sense for the developer documentation
to be publically available. Since the repo will be available on Github, it makes sense
to also host the developer documentation there.

This commit adds a workflow to publish the mkdocs build as a Github page.

The Github page needs to be enabled in the Github settings of the repo.

For more info see:
https://github.com/mhausenblas/mkdocs-deploy-gh-pages\#building-with-github_token