Move local build instructions to docs from README
To allow referencing the setup instructions in the docs, they should be
included in the docs. It is undesirable to have this documented in two
places though.

Once the docs are publicly available it is not necessary create a local
setup to then read the documentation. This means the setup instructions
can be moved from the README to the docs without becoming unavailable.