Add setup instructions back to README
The local setup is one of the first points of information that a new
developer on the project is going to be interested. This should not be
hidden away in the docs.

To also have this information available in the docs (without having
to refer back to the README) it is now included in both places. This is
an undesirable duplication, but I could not come up with a way to avoid
it while providing the necessary information where needed.

These instructions are also unlikely to change, making the additional
overhead small.