Add setting for boost reduction factor
Before, the reduction factor was hard coded in the mixin.

The hard coding was not ideal because the exact value necessary is not
clear at this point.

Now, a setting was added that pulls the reduction factor from an
environment variable (but defaults to 0.5). Having this setting be
configurable through an environment variable allows to adjust the search
ranking on the server without the need to change code. This should make
it easier to run experiments with this value in production or staging.