Rename boost factor setting
`SEARCH_BOOST_REDUCTION_FACTOR_NEWS_PAGE` seems to suggest that higher
values mean higher reduction. That is not the case. It is only a factor
that is applied to the relevance score. This seems to be better
expressed by `SEARCH_BOOST_FACTOR_NEWS_PAGE`. The fact that this value
needs to be between 0 and 1 is documented in a comment and the docs.