Add navigation setting for header title
Previously, the header title used in the FIS header was hardcoded in the
template.

This prevent reuse of the templates with this header for other micro
sites.

Now, a new navigation setting is added to allow to set the header title
site wide, but choose different values for different sites.

This change requires the FIS page on staging and production to be
updated and for this setting to be added.