Prevent auth headers from being deleted
We updated `django-basic-auth-ip-whitelist` (baipw) to the master
version directly on GitHub to include a fix that prevents 401 responses
from being cached. The previous version was 0.2.1.

After this update, it was noticed that Wagtail preview does not work
anymore. It keeps requesting the basic auth credentials, even if the
correct ones had been entered.

This issue was apparently caused by a change in baipw, that was
introduced to allow interaction with Django REST Framework. This change
meant that the authorization headers would be deleted after being
checked by the baipw middleware.

Luckily, a setting was added that allow disabling this deletion, which
fixes previews in Wagtail.

This commit adds this setting to prevent the deletion of the
authorization headers.

See also:
https://gitlab.com/tmkn/django-basic-auth-ip-whitelist/-/issues/11
https://gitlab.com/tmkn/django-basic-auth-ip-whitelist/-/issues/13
https://gitlab.com/tmkn/django-basic-auth-ip-whitelist/-/merge_requests/17