Move OGL and credit out of footer nav container
Per design, the OGL statement and credit/copyright line are always below
the navigation elements (site and social links).

Before, the OGL and the credit elements where included in the navigation
container. On large displays, this lead to the OGL element being
displayed on the same line as the links.

Now, the OGL and credit elements are moved out of the nav container.
This way, they are always displayed below the navigational elements.
This seems closer to the designs.

See also:
https://zeroheight.com/5ef2b1a81/p/799c94-footers/b/17dca6