Add search input help text to template
The search input help text might be empty. Therefore, I did not want
to change any of the margins on the form itself.

But, I also wanted the help text closer to the input then the results.
This is why I am using a negative margin on the helptext to move it
closer to the input.

The bottom margins are the same as the form itself is using.