Redirect from campaign index page on serve
The campaign index page is only structural an not supposed to have any content.
There won't be any links to this site, but somebody could get to to the site
by manually changing the URL.

The serve method is overriden to always return a redirect from the campaign index page
to the root page of the current site.