Flexible section content
Before, the sections content was limited to a list of section content
blocks (image or embed, headline, paragraph). Additionally, each section
could have a full width or directory banner.

This was not the intent of the design. Each element should be able to
appear at any place in the section (content, full width banner,
directory banner). This is now enabled through a stream block.