Use wagtail factories for standard pages
Before, these were using the normal Django model factories. This would
require the path fields etc to be set manually when the model is created
in the data base (e.g. the factory method `create` is used to save the model).

This can be simplified by using the `wagtail_factories` package, which was
already present in the project.