Use old style TableBlock in Accordion
For release 50.0, TableBlock was converted to a TypedTableBlock, nested
inside CaptionedTableBlock (a StructBlock). A data migration (d511b5b73)
converted all table blocks which were top-level children of the field,
but missed those table blocks nested in Accordions.

Because the data structure did not match what CaptionedTableBlock
expected, these blocks were not editable, and did not display on the
front-end.

This commit changes nested table blocks to TableBlock, while keeping
top-level ones, whose data has been migrated, as the newer format.

In future, we should revert this, and add a data migration for tables
inside Accordions. For now, this restores those blocks to be editable
and renderable.