Revert bc649e8e and 381a1535
This commit removes the custom validation to ensure
that a RecruitmentHomePage has a related_recruitment_index_page
specified.

This is because this presents a chicken-egg problem, where, the
RecruitmentIndexPage cannot be created anywhere else unless under
a RecruitmentHomePage, which means the latter must exist prior to
the instantiation of the former.